System AddOns eingefhrt. Feste AddOns in REDAXO. Werden zwar wie normale AddOns verwendet, sind aber so wichtig das diese immer vorhanden sein sollen.

+++ redaxo3_0/redaxo/include/pages/addon.inc.php
@@ -3,7 +3,7 @@
 /** 
  *  
  * @package redaxo3 
- * @version $Id: addon.inc.php,v 1.17 2007/06/07 10:56:57 kristinus Exp $ 
+ * @version $Id: addon.inc.php,v 1.18 2007/06/07 11:20:55 kristinus Exp $ 
  */
 
 include_once $REX['INCLUDE_PATH']."/functions/function_rex_other.inc.php";
@@ -121,7 +121,13 @@ if ($SP)
 
   foreach ($ADDONS as $cur)
   {
-    $delete = '<a href="index.php?page=addon&amp;addonname='.$cur.'&amp;delete=1" onclick="return confirm(\''.$I18N->msg('delete').' ?\')">'.$I18N->msg("addon_delete").'</a>';
+  	if (!isset($REX['ADDON']['system'][$cur]) || !$REX['ADDON']['system'][$cur])
+  	{
+  		$delete = '<a href="index.php?page=addon&amp;addonname='.$cur.'&amp;delete=1" onclick="return confirm(\''.$I18N->msg('delete').' ?\')">'.$I18N->msg("addon_delete").'</a>';
+  	}else
+  	{
+  		$delete = ''.$I18N->msg("addon_systemaddon").'';
+  	}
 
     if ($REX['ADDON']['install'][$cur] == 1)
     {
